requirements.txt
fastapi==0.111.0
httpx==0.27.0
pydantic==1.10.9
PyJWT==2.8.0
PyJWT==1.7.1
PyJWT==2.8.0
typing_extensions==4.12.2
